ai_service/pyproject.toml
[project]
name = "ai-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.123.7",
    "langgraph>=1.0.5",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "uvicorn>=0.38.0",
]
